third_party/blink/web_tests/fast/dom/shadow/style-sharing-sibling-shadow.html

<!DOCTYPE html>

<style>
    #sandbox div {
        width: 50px;
        height: 50px;
    }
</style>

<script src="../../../resources/js-test.js"></script>

<div id="sandbox">
    <div></div>
    <div></div>
</div>

<script>
description("Siblings should only share if their host rules match");

document.body.offsetTop;
host = document.getElementById("sandbox").firstElementChild;
host.attachShadow({mode: 'open'}).innerHTML = "<style>:host { background: red; }</style>";
shouldBeEqualToString("getComputedStyle(host).backgroundColor", "rgb(255, 0, 0)");
</script>
