third_party/blink/web_tests/fast/dom/shadow/style-sharing-sibling-shadow-expected.txt

Siblings should only share if their host rules match

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(host).backgroundColor is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
